+++ src/main/java/com/ericsson/eiffel/semantics/events/EiffelAlertAcknowledgedEvent.java
@@ -19,9 +19,7 @@ import java.util.List;
 import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
 
-public class EiffelAlertAcknowledgedEvent
-    extends Event
-{
+public class EiffelAlertAcknowledgedEvent extends Event {
 
     /**
      *