+++ src/main/java/com/ericsson/eiffel/semantics/events/EiffelArtifactDeployedEvent.java
@@ -19,9 +19,7 @@ import java.util.List;
 import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
 
-public class EiffelArtifactDeployedEvent
-    extends Event
-{
+public class EiffelArtifactDeployedEvent extends Event {
 
     /**
      *