+++ src/main/java/com/ericsson/eiffel/semantics/events/EiffelFlowContextDefinedEvent.java
@@ -19,9 +19,7 @@ import java.util.List;
 import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
 
-public class EiffelFlowContextDefinedEvent
-    extends Event
-{
+public class EiffelFlowContextDefinedEvent extends Event {
 
     /**
      *