+++ src/main/java/com/ericsson/eiffel/semantics/events/EiffelTestCaseTriggeredEventMeta.java
@@ -21,8 +21,7 @@ import java.util.Map;
 import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
 
-public class EiffelTestCaseTriggeredEventMeta implements Meta
-{
+public class EiffelTestCaseTriggeredEventMeta implements Meta {
 
     /**
      * 
@@ -47,7 +46,8 @@ public class EiffelTestCaseTriggeredEventMeta implements Meta
      */
     @SerializedName("version")
     @Expose
-    private EiffelTestCaseTriggeredEventMeta.Version version = EiffelTestCaseTriggeredEventMeta.Version.fromValue("1.1.0");
+    private EiffelTestCaseTriggeredEventMeta.Version version = EiffelTestCaseTriggeredEventMeta.Version
+            .fromValue("3.0.0");
     /**
      * 
      * (Required)
@@ -170,7 +170,7 @@ public class EiffelTestCaseTriggeredEventMeta implements Meta
         private final static Map<String, EiffelTestCaseTriggeredEventMeta.Type> CONSTANTS = new HashMap<String, EiffelTestCaseTriggeredEventMeta.Type>();
 
         static {
-            for (EiffelTestCaseTriggeredEventMeta.Type c: values()) {
+            for (EiffelTestCaseTriggeredEventMeta.Type c : values()) {
                 CONSTANTS.put(c.value, c);
             }
         }
@@ -201,13 +201,13 @@ public class EiffelTestCaseTriggeredEventMeta implements Meta
 
     public enum Version {
 
-        @SerializedName("1.1.0")
-        _1_1_0("1.1.0");
+        @SerializedName("3.0.0")
+        _3_0_0("3.0.0");
         private final String value;
         private final static Map<String, EiffelTestCaseTriggeredEventMeta.Version> CONSTANTS = new HashMap<String, EiffelTestCaseTriggeredEventMeta.Version>();
 
         static {
-            for (EiffelTestCaseTriggeredEventMeta.Version c: values()) {
+            for (EiffelTestCaseTriggeredEventMeta.Version c : values()) {
                 CONSTANTS.put(c.value, c);
             }
         }