+++ src/main/java/com/ericsson/eiffel/semantics/events/EiffelTestExecutionRecipeCollectionCreatedEventMeta.java
@@ -21,8 +21,7 @@ import java.util.Map;
 import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
 
-public class EiffelTestExecutionRecipeCollectionCreatedEventMeta implements Meta
-{
+public class EiffelTestExecutionRecipeCollectionCreatedEventMeta implements Meta {
 
     /**
      * 
@@ -47,7 +46,8 @@ public class EiffelTestExecutionRecipeCollectionCreatedEventMeta implements Meta
      */
     @SerializedName("version")
     @Expose
-    private EiffelTestExecutionRecipeCollectionCreatedEventMeta.Version version = EiffelTestExecutionRecipeCollectionCreatedEventMeta.Version.fromValue("2.1.0");
+    private EiffelTestExecutionRecipeCollectionCreatedEventMeta.Version version = EiffelTestExecutionRecipeCollectionCreatedEventMeta.Version
+            .fromValue("4.0.0");
     /**
      * 
      * (Required)
@@ -170,7 +170,7 @@ public class EiffelTestExecutionRecipeCollectionCreatedEventMeta implements Meta
         private final static Map<String, EiffelTestExecutionRecipeCollectionCreatedEventMeta.Type> CONSTANTS = new HashMap<String, EiffelTestExecutionRecipeCollectionCreatedEventMeta.Type>();
 
         static {
-            for (EiffelTestExecutionRecipeCollectionCreatedEventMeta.Type c: values()) {
+            for (EiffelTestExecutionRecipeCollectionCreatedEventMeta.Type c : values()) {
                 CONSTANTS.put(c.value, c);
             }
         }
@@ -201,13 +201,13 @@ public class EiffelTestExecutionRecipeCollectionCreatedEventMeta implements Meta
 
     public enum Version {
 
-        @SerializedName("2.1.0")
-        _2_1_0("2.1.0");
+        @SerializedName("4.0.0")
+        _4_0_0("4.0.0");
         private final String value;
         private final static Map<String, EiffelTestExecutionRecipeCollectionCreatedEventMeta.Version> CONSTANTS = new HashMap<String, EiffelTestExecutionRecipeCollectionCreatedEventMeta.Version>();
 
         static {
-            for (EiffelTestExecutionRecipeCollectionCreatedEventMeta.Version c: values()) {
+            for (EiffelTestExecutionRecipeCollectionCreatedEventMeta.Version c : values()) {
                 CONSTANTS.put(c.value, c);
             }
         }